requirements.txt
python-dotenv
paho-mqtt 
RPi.GPIO